Changed top left to top right.
There was an issue in most cases where the logo would cover up a couple names in the group menu in the top left preventing players from seeing their names. I have moved this to the top right as there is never anything in the HUD that the logo can cover while it's up there.