Added Link + href example.
I have added a link to the ArmA 3 wiki that takes people to the Structured Text page. This will help them be able to use this script better and have no problems creating links etc. In the app to rules pages and such.

I have also included a href example at the bottom of the script so it adds a link example at the same time as crediting the Add-on. ;)